+++ src/contracts/conversion/image_traindata_conveter.py
@@ -1,0 +1,12 @@
+from abc import ABC, abstractmethod
+
+
+class ImageConverterAbstract(ABC):
+
+    @abstractmethod
+    def convert(self, image):
+        pass
+
+    @abstractmethod
+    def convert(self, image, annotations):
+        pass
