+++ src/contracts/providing/logging_path_provider.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class LoggingPathProviderAbstract(ABC):
+
+    @staticmethod
+    @abstractmethod
+    def get_path():
+        pass
