--- src/dataset_augmentations/label_augmentation.py
@@ -1,1 +1,0 @@
-#Здесь будут аугментировать данные из json, чтобы метки соответствовали изображениям
